Organized files into packages, created the Menu for all zooming features (still need to determine how to pass a page instance), updated Zoomer.interact again to reflect changes in input from MenuBar, deleted some obsolete files
