+++ .idea/compiler.xml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="CompilerConfiguration">
-    <bytecodeTargetLevel target="11" />
+    <bytecodeTargetLevel target="11">
+      <module name="projects.course-project-team-96-1.5note-norman-edition" target="17" />
+    </bytecodeTargetLevel>
   </component>
 </project>
