+++ build.gradle
@@ -8,13 +8,14 @@ repositories {
 }
 
 dependencies {
+    implementation 'org.xerial:sqlite-jdbc:3.39.4.0'
     implementation 'junit:junit:4.13.1'
     testImplementation('org.junit.jupiter:junit-jupiter:5.6.0')
 }
 
 javafx {
-    version = "17"
-    modules = [ 'javafx.controls' ]
+    version = "19"
+    modules = [ 'javafx.controls', 'javafx.media' ]
 }
 
 application {
@@ -26,3 +27,18 @@ test {
         events "passed", "skipped", "failed"
     }
 }
+
+jar {
+    archiveFileName = 'Notebook.jar';
+
+    duplicatesStrategy = 'exclude';
+
+    manifest {
+        attributes "Main-Class": "Main"
+    }
+
+    from {
+        configurations.runtimeClasspath.collect { it.isDirectory() ? it : zipTree(it) }
+    }
+}
+
