+++ res/css/fonts.css
@@ -1,0 +1,4 @@
+@font-face {
+    font-family: Cantarell;
+    src: url("../font/cantarell/Cantarell-VF.otf");
+}
