+++ res/css/page.css
@@ -1,0 +1,3 @@
+.page {
+    -fx-background-color: white;
+}
