+++ res/css/toolbar.css
@@ -1,0 +1,5 @@
+.toolbar {
+    -fx-background-color: -fx-background;
+    -fx-border-color: -fx-box-border;
+    -fx-border-width: 0 0 1 0;
+}
