--- src/main/java/SearchBarController.java
@@ -1,11 +1,0 @@
-public class SearchBarController {
-
-    // Refine return type. As of right now, documentation implies it can be a
-    // media, file, or subsection of text in a text box???
-    // I have started adjusting the CRC cards so that they only implies returning the text media,
-    // But tags are still an issue.
-    public String search(String input) {
-        Searcher searcher = new Searcher();
-        return searcher.search(input);
-    }
-}
