+++ src/main/java/app/controllers/SearchBarController.java
@@ -1,0 +1,15 @@
+package app.controllers;
+
+import app.interaction_managers.Searcher;
+
+public class SearchBarController {
+
+    // Refine return type. As of right now, documentation implies it can be a
+    // media, file, or subsection of text in a text box???
+    // I have started adjusting the CRC cards so that they only implies returning the text media,
+    // But tags are still an issue.
+    public String search(String input) {
+        Searcher searcher = new Searcher();
+        return searcher.search(input);
+    }
+}
