+++ src/main/java/app/controllers/Zoomable.java
@@ -1,0 +1,7 @@
+package app.controllers;
+
+/** interface between controller level and app.gui.page.Page (in GUI level) */
+// TODO change this comment
+
+public interface Zoomable {
+}
