+++ src/main/java/app/controllers/ZoomerController.java
@@ -1,0 +1,9 @@
+package app.controllers;
+
+//class to pass zoom percentage back to app.gui.page.Page
+public class ZoomerController implements app.interaction_managers.ZoomerControllerInt {
+    @Override
+    public void zoom(double zoomFactor) {
+
+    }
+}
