+++ src/main/java/app/interaction_managers/InteractionManager.java
@@ -1,4 +1,4 @@
-// TODO temp
+package app.interaction_managers;// TODO temp
 import javafx.scene.Node;
 
 public interface InteractionManager {
