+++ src/main/java/app/interaction_managers/ZoomerControllerInt.java
@@ -1,0 +1,8 @@
+package app.interaction_managers;
+
+/** interface between app.interaction_managers.Zoomer and app.controllers.ZoomerController */
+// TODO change this comment
+
+public interface ZoomerControllerInt {
+    public void zoom(double zoomFactor);
+}
