+++ src/main/java/app/media_managers/ImageModifier.java
@@ -1,4 +1,6 @@
-public class ImageModifier implements MediaManager{
+package app.media_managers;
+
+public class ImageModifier implements MediaManager {
     private String caption;
 
     public void createObject() {
