+++ src/main/java/app/media_managers/MediaManager.java
@@ -1,3 +1,5 @@
+package app.media_managers;
+
 interface MediaManager {
     public void addMedia();
     public void modifyMedia();
